feat(platform): cursor placement; move getBookCodeFromUsj to shared
- Move getBookCodeFromUsj from scripture-utilities into libs/shared utils
  and drop it from the published utilities API surface.
- Adjust ScriptureReferencePlugin, Editor, and USJ node helpers for PT-3208.
- Regenerate scripture-utilities API report.

Made-with: Cursor
