+++ libs/shared/src/utils/usj/index.ts
@@ -1,1 +1,2 @@
 export * from "./contentToLexicalNode.js";
+export * from "./getBookCodeFromUsj.js";
